cost function refactor
