+++ nodejs-server/service/DacsService.js
@@ -1,6 +1,6 @@
 'use strict';
 const Summary = require('../models/summary');
-const helpers = require('./helpers')
+const helpers = require('../helpers/helpers')
 
 /**
  * Provides summary data for each data assembly center.
