+++ nodejs-server/service/GridService.js
@@ -1,6 +1,6 @@
 'use strict';
 const Grid = require('../models/grid');
-const helpers = require('./helpers')
+const helpers = require('../helpers/helpers')
 const GJV = require('geojson-validation');
 const geojsonArea = require('@mapbox/geojson-area');
